mycelium/canary_server/requirements.txt
aiohttp==3.9.5
asyncpg==0.29.0
nats-py==2.7.2
httpx==0.27.0
structlog==24.1.0
